감속 비율 변경 기능 추가

+++ LibNurirobotV00/Struct/NuriRatio.cs
@@ -8,7 +8,7 @@ namespace LibNurirobotV00.Struct
         /// <summary>
         /// 감속비
         /// </summary>
-        public float Ratio { get; set; }
+        public decimal Ratio { get; set; }
         /// <summary>
         /// 피드백 모드
         /// </summary>
